+++ src/StitchPreview.ts
@@ -1,13 +1,14 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
+import axios from 'axios';
 import { StitchView } from './StitchView';
 import { FileScrambler } from './FileScrambler';
 import { COMMANDS, CONSTANTS, MESSAGES } from './constants';
-import { PreviewContext, ScenarioSource, StitchResponse, TreeItem } from './types';
-import axios from 'axios';
+import { CommandAction, ICommand, PreviewContext, RenderTemplateStepResult, ScenarioSource, StitchResponse, TreeItem } from './types';
+import { PdfPreview } from './PdfPreview';
 
 export class StitchPreview {
-	
+
     public static currentPreview: StitchPreview | undefined;
 
     private _disposables: vscode.Disposable[] = [];
@@ -28,17 +29,18 @@ export class StitchPreview {
             return;
         }
 
-	    const endpoint = vscode.workspace.getConfiguration().get<string>(CONSTANTS.configKeyEndpointUrl);
+        const endpoint = vscode.workspace.getConfiguration().get<string>(CONSTANTS.configKeyEndpointUrl);
         if (!endpoint) {
             vscode.window.showErrorMessage(MESSAGES.endpointUrlNotConfigured);
             return;
         }
 
         const showOptions = {
-            viewColumn: vscode.ViewColumn.Two, 
+            viewColumn: vscode.ViewColumn.Two,
             preserveFocus: true
         };
         const options = {
+            enableScripts: true,
             localResourceRoots: [vscode.Uri.file(path.join(extensionUri.path, 'assets'))]
         };
         const panel = vscode.window.createWebviewPanel('stitchPreview', '', showOptions, options);
@@ -77,7 +79,7 @@ export class StitchPreview {
         });
     }
 
-    public static currentResponse() : StitchResponse | undefined {
+    public static currentResponse(): StitchResponse | undefined {
         const current = StitchPreview.currentPreview;
         if (!current) { return; }
 
@@ -85,22 +87,75 @@ export class StitchPreview {
     }
 
     public static async openScenarioFile(treeItem: TreeItem): Promise<void> {
-		const current = StitchPreview.currentPreview;
+        const current = StitchPreview.currentPreview;
         if (!current || !current._scenario) { return; }
 
         if (treeItem.path === 'Model') {
             const uri = vscode.Uri.file(path.join(current._scenario.path, 'input.txt'));
             await vscode.window.showTextDocument(uri);
             return;
-        } 
-        
+        }
+
         const match = treeItem.path.match(/Steps.([a-zA-Z0-9_-]+).Model/);
         const stepName = match && match[1];
         if (stepName) {
             const uri = vscode.Uri.file(path.join(current._scenario.path, `step.${stepName}.txt`));
             await vscode.window.showTextDocument(uri);
         }
-	}	
+    }
+
+    public static handleCommand(command: ICommand, extensionUri: vscode.Uri) {
+        const response = StitchPreview.currentResponse();
+        if (!response) { return; }
+
+        switch (command.action) {
+            case CommandAction.viewStepRequest:
+                const step = command.content;
+                this.showRendered({
+                    filename: `stitch-step-request-${step}`,
+                    content: response.requests[step].content.trim()
+                });
+                return;
+            case CommandAction.viewStepResponse:
+                if (response.integrationContext.steps[command.content]?.$type !== CONSTANTS.renderTemplateStepResultType) { return; }
+                const renderResponse = <RenderTemplateStepResult>response.integrationContext.steps[command.content];
+                if (renderResponse.response.contentType !== 'application/pdf') { return; }
+                PdfPreview.createOrShow(command.content ,extensionUri);
+                PdfPreview.setOrUpdatePdfData(command.content, renderResponse.response.content);
+                return;
+            case CommandAction.viewIntegrationResponse:
+                this.showRendered({
+                    filename: `stitch-response`,
+                    content: JSON.stringify(response.result, null, 2)
+                });
+                return;
+        }
+    }
+
+    static showRendered(options: { filename: string; content: string; }) {
+        if (!options.content) { return; }
+
+        const firstChar = options.content[0];
+        const ext = firstChar === '<' ? '.xml' : firstChar === '{' ? '.json' : '.txt';
+
+        const untitledFile = vscode.Uri.parse(`untitled:${options.filename}${ext}`);
+        this.updateRendered(untitledFile, options.content, true);
+    }
+
+    static updateRendered(untitledUri: vscode.Uri, content: string, show: boolean = false) {
+        vscode.workspace.openTextDocument(untitledUri).then(document => {
+            const lastLine = document.lineAt(document.lineCount-1);
+            const edit = new vscode.WorkspaceEdit();
+            edit.replace(untitledUri, new vscode.Range(new vscode.Position(0,0), lastLine.range.end), content);
+            return vscode.workspace.applyEdit(edit).then(success => {
+                if (success && show) {
+                    vscode.window.showTextDocument(document, undefined, true);
+                } else {
+                    vscode.window.showInformationMessage('Error!');
+                }
+            });
+        });
+    }
 
     constructor(
         private _panel: vscode.WebviewPanel,
@@ -119,7 +174,7 @@ export class StitchPreview {
         vscode.window.onDidChangeActiveTextEditor((e): void => {
             e && ['file'].includes(e.document.uri.scheme) && this._update(e);
         }, null, this._disposables);
-        
+
         vscode.workspace.onDidChangeConfiguration(e => {
             if (e.affectsConfiguration(CONSTANTS.configKeyEndpointUrl)) {
                 this._onUpdateEndoint();
@@ -133,6 +188,7 @@ export class StitchPreview {
         vscode.commands.executeCommand('setContext', CONSTANTS.previewActiveContextKey, false);
         this._panel.dispose();
         this._statusBar.dispose();
+        PdfPreview.disposeAll();
 
         while (this._disposables.length) {
             const x = this._disposables.pop();
@@ -153,7 +209,7 @@ export class StitchPreview {
         this._testEndpoint = `${endpoint}/test`;
         vscode.window.showInformationMessage('The Stitch test endpoint has been updated to: ' + endpoint);
         this._update();
-	}
+    }
 
     private _getContext(textEditor?: vscode.TextEditor): PreviewContext | undefined {
         const activeEditor = textEditor || vscode.window.activeTextEditor;
@@ -183,17 +239,18 @@ export class StitchPreview {
     private _update(textEditor?: vscode.TextEditor) {
 
         const integrationContext = this._getContext(textEditor);
-        if (!integrationContext) { 
+        if (!integrationContext) {
             this._view.displayError({
                 title: `No ${CONSTANTS.integrationExtension} file found`,
                 description: `Please open an *${CONSTANTS.integrationExtension} file or directory to enable the preview!`
             });
             vscode.commands.executeCommand('setContext', CONSTANTS.previewActiveContextKey, false);
-            return; 
+            return;
         }
 
         if (this._isContextChanged(integrationContext)) {
             this._scenario = undefined;
+            PdfPreview.disposeAll();
         }
         this._context = integrationContext;
         this._panel.title = `${CONSTANTS.panelTitlePrefix}${integrationContext.integrationFilename}`;
@@ -228,6 +285,8 @@ export class StitchPreview {
                 this._view.displayResult(res.data, integrationContext, scenario);
                 if (res.data.result) {
                     this._result = <StitchResponse>res.data;
+                    this._updateRenderedUntitled();
+                    this._updateRenderedPdf();
                     vscode.commands.executeCommand(COMMANDS.responseUpdated); // so other Components know to request the latest response
                 }
             })
@@ -238,15 +297,51 @@ export class StitchPreview {
                 });
             });
     }
+
+    private _updateRenderedUntitled() {
+
+        const open = vscode.workspace.textDocuments.filter(t => t.uri.scheme === 'untitled' && t.fileName.startsWith('stitch-'));
+        const response = this._result;
+        if (open.length === 0 || !response) {return;}
+
+        open.forEach(o => {
+            if (o.fileName.startsWith('stitch-response.')) {
+                StitchPreview.updateRendered(o.uri, JSON.stringify(response.result, null, 2));
+            } else {
+                let match = o.fileName.match(/^stitch-step-request-(.*?)\./);
+                if (match?.length === 2) {
+                    StitchPreview.updateRendered(o.uri, response.requests[match[1]].content.trim());
+                }
+            }
+        });
+    }
+
+    private _updateRenderedPdf() {
+        if (!this._result) { return; }
+        const response = this._result;
+
+        for (const stepId of PdfPreview.renderedSteps) {
+            if (response.integrationContext.steps[stepId]?.$type !== CONSTANTS.renderTemplateStepResultType) { 
+                PdfPreview.disposeRenderedStep(stepId);
+                continue;
+            }
+            const renderResponse = <RenderTemplateStepResult>response.integrationContext.steps[stepId];
+            if (renderResponse.response.contentType !== 'application/pdf') { 
+                PdfPreview.disposeRenderedStep(stepId);
+                continue;
+             }
+            PdfPreview.setOrUpdatePdfData(stepId, renderResponse.response.content);
+        }
+    }
+
     private _isContextChanged(newContext: PreviewContext) {
         return this._context?.integrationFilePath !== newContext.integrationFilePath;
     }
 
-    private _readWorkspaceFile(filepath: string) : string | undefined {
+    private _readWorkspaceFile(filepath: string): string | undefined {
         const textDoc = vscode.workspace.textDocuments.find(doc => doc.fileName === filepath);
         if (textDoc) {
             return textDoc.getText();
         }
     }
 }
-