+++ src/StitchView.ts
@@ -1,12 +1,20 @@
 import * as vscode from 'vscode';
-import { HttpStepResult, PreviewContext, ScenarioSource, StepRequest, StepResult, StitchError, StitchResponse } from './types';
+import { CONSTANTS } from './constants';
+import { StitchPreview } from './StitchPreview';
+import { CommandAction, HttpStepResult, ICommand, PreviewContext, RenderTemplateStepResult, ScenarioSource, StepRequest, StepResult, StitchError, StitchResponse } from './types';
 
 export class StitchView {
 
     private _stylesMainUri: vscode.Uri;
+    private _disposables: vscode.Disposable[] = [];
 
     constructor(private _webview: vscode.Webview, extensionUri: vscode.Uri) {
         this._stylesMainUri = _webview.asWebviewUri(vscode.Uri.joinPath(extensionUri, 'assets', 'style.css'));
+        _webview.onDidReceiveMessage(
+            (command: ICommand) => { StitchPreview.handleCommand(command, extensionUri); },
+            undefined,
+            this._disposables
+        );
     }
 
     public displayError(error: StitchError, extraBody?: string): void {
@@ -19,7 +27,7 @@ export class StitchView {
         const contextHtml = this._getContextHtml(previewContext, scenario);
 
         if (!data.result) {
-            if (data.ClassName ==='Core.Exceptions.StitchResponseSerializationException') {
+            if (data.ClassName === 'Core.Exceptions.StitchResponseSerializationException') {
                 return this.displayError({
                     title: data.Message,
                     description: `<pre><code>${data.ResultBody}</code></pre><strong>Exception:</strong><br />${data.InnerException?.Message}`
@@ -41,6 +49,7 @@ export class StitchView {
         var resultStatusCode = response.resultStatusCode ? response.resultStatusCode : 200;
 
         const htmlBody = `<h1>Output</h1>
+                          <button onclick="vscode.postMessage({action: ${CommandAction.viewIntegrationResponse} });">view as file</button>
                           <pre><code>${JSON.stringify(response.result, null, 2)}</code></pre>
                           <p>Statuscode: ${resultStatusCode}</p>
                           ${contextHtml}
@@ -62,21 +71,54 @@ export class StitchView {
                 </table>`;
     }
 
-    private _getStepHtml(stepId: string, step: StepResult | HttpStepResult, requests: Record<string, StepRequest>) {
-        
-        if (step.$type !== 'Core.Entities.StepResults.HttpStepResult, Core') {
+    private _getStepHtml(stepId: string, step: StepResult, requests: Record<string, StepRequest>) {
+        if (step.$type === CONSTANTS.httpStepResultTypeType) {
+            const httpStep = <HttpStepResult>step;
+            return `<div>
+                        <h4>${stepId}</h4>
+                        <p>${httpStep.request.method} - ${httpStep.request.url}</p>
+                        <button onclick="vscode.postMessage({action: ${CommandAction.viewStepRequest}, content: '${stepId}' });">view as file</button>
+                        <pre><code>${this._escapeHtml(requests[stepId].content)}</code></pre>
+                    </div>`;
+        }
+        else if (step.$type === CONSTANTS.renderTemplateStepResultType) {
+            const renderStep = <RenderTemplateStepResult>step;
+            var stepHtml = `<div>
+                                <h4>${stepId}</h4>
+                                <h5>INPUT | Zip (base64)</h5>
+                                <pre><code>${StitchView._trimQuotationMarks(requests[stepId].content)}</code></pre>`;
+            if (renderStep.response.isSuccessStatusCode) {
+                stepHtml +=    `<h5>OUTPUT | Pdf (base64)</h5>
+                                <button onclick="vscode.postMessage({action: ${CommandAction.viewStepResponse}, content: '${stepId}' });">view as file</button>
+                                <pre><code>${renderStep.response.content}</code></pre>`;
+            }
+            else {
+                stepHtml +=    `<h5>OUTPUT | Error (${renderStep.response.statusCode})
+                                <p>${renderStep.response.errorMessage}</p>`;
+            }
+            stepHtml += '</div>';
+            return stepHtml;
+        }
+        else {
             return `<div>
                         <h4>${stepId}</h4>
                         <p>Started: ${step.started}</p>
                     </div>`;
         }
         
-        const httpStep = <HttpStepResult>step;
-        return `<div>
-                    <h4>${stepId}</h4>
-                    <p>${httpStep.request.method} - ${httpStep.request.url}</p>
-                    <pre><code>${this._escapeHtml(requests[stepId].content)}</code></pre>
-                </div>`;
+
+    }
+
+    private static _trimQuotationMarks(untrimmed: string) {
+        let result = untrimmed;
+        if (untrimmed.startsWith('"')) {
+            result = result.substr(1);
+        }
+        if (untrimmed.endsWith('"')) {
+            result = result.slice(0, -1);
+        }
+
+        return result;
     }
 
     private _escapeHtml(unsafe: string) {
@@ -94,12 +136,25 @@ export class StitchView {
             <html lang="en">
             <head>
                 <meta charset="UTF-8">
-                <meta http-equiv="Content-Security-Policy" content="default-src 'none'; style-src ${cspSource}; img-src ${cspSource} https:;">
                 <meta name="viewport" content="width=device-width, initial-scale=1.0">
-				<link href="${this._stylesMainUri}" rel="stylesheet">
                 <title>Stitch Preview</title>				
+
+                <meta http-equiv="Content-Security-Policy" content="
+                    default-src 'none';
+                    style-src ${cspSource};
+                    script-src 'unsafe-inline';
+                    img-src ${cspSource} https:;">
+                <link href="${this._stylesMainUri}" rel="stylesheet">
+                <script>
+                    window.acquireVsCodeApi = acquireVsCodeApi;
+                </script>
             </head>
-            <body>${htmlBody}</body>
+            <body>
+                ${htmlBody}
+                <script>
+                    const vscode = window.acquireVsCodeApi();
+                </script>
+            </body>
             </html>`;
     }
 }